LJH/NER BERT-BiLSTM-CRF/requirements.txt
absl-py==0.15.0
amqp==5.0.6
apache-libcloud==3.4.0
asgiref==3.4.1
astunparse==1.6.3
auto-labeling-pipeline==0.1.21
billiard==3.6.4.0
boto3==1.20.5
botocore==1.23.5
cachetools==4.2.4
celery==5.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
clang==5.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
colour==0.1.5
conllu==4.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.11
Django==3.2.9
django-celery-results==2.2.0
django-cors-headers==3.10.0
django-drf-filepond==0.4.0
django-filter==21.1
django-polymorphic==3.0.0
django-rest-polymorphic==0.1.9
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-csv==2.1.1
djangorestframework-xml==2.0.0
doccano==1.5.5
drf-yasg==1.20.0
ecdsa==0.17.0
environs==9.3.4
et-xmlfile==1.1.0
filelock==3.3.1
flatbuffers==1.12
furl==2.1.3
gast==0.4.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.41.0
gunicorn==20.1.0
h5py==3.1.0
huggingface-hub==0.0.19
idna==3.3
inflection==0.5.1
itypes==1.2.0
jieba==0.42.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kombu==5.2.1
lml==0.1.0
Markdown==3.3.4
MarkupSafe==2.0.1
marshmallow==3.14.0
numpy==1.19.5
oauthlib==3.1.1
openpyxl==3.0.9
opt-einsum==3.3.0
orderedmultidict==1.0.1
packaging==21.0
pandas==1.3.4
prompt-toolkit==3.0.22
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
pyexcel==0.6.7
pyexcel-io==0.6.5
pyexcel-xlsx==0.6.0
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
regex==2021.10.21
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.2
seqeval==1.2.2
shortuuid==1.0.8
six==1.15.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
SQLAlchemy==1.4.27
sqlparse==0.4.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-addons==0.14.0
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
texttable==1.6.4
threadpoolctl==3.0.0
tokenizers==0.10.3
tqdm==4.62.3
transformers==4.11.3
typeguard==2.13.0
typing-extensions==3.7.4.3
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
waitress==2.0.0
wcwidth==0.2.5
Werkzeug==2.0.2
whitenoise==5.3.0
wincertstore==0.2
wrapt==1.12.1
